+++ src/Phan/Analysis/PreOrderAnalysisVisitor.php
@@ -11,6 +11,7 @@ use Phan\CodeBase;
 use Phan\Config;
 use Phan\Exception\CodeBaseException;
 use Phan\Exception\NodeException;
+use Phan\Exception\RecursionDepthException;
 use Phan\Exception\UnanalyzableException;
 use Phan\Issue;
 use Phan\IssueFixSuggester;
@@ -526,10 +527,14 @@ class PreOrderAnalysisVisitor extends ScopeVisitor
         }
         if (!$func->isReturnTypeUndefined()) {
             $func_return_type = $func->getUnionType();
-            $func_return_type_can_cast = $func_return_type->canCastToExpandedUnionType(
-                Type::fromNamespaceAndName('\\', 'Generator', false)->asUnionType(),
-                $this->code_base
-            );
+            try {
+                $func_return_type_can_cast = $func_return_type->canCastToExpandedUnionType(
+                    Type::fromNamespaceAndName('\\', 'Generator', false)->asUnionType(),
+                    $this->code_base
+                );
+            } catch (RecursionDepthException $_) {
+                return;
+            }
             if (!$func_return_type_can_cast) {
                 // At least one of the documented return types must
                 // be Generator, Iterable, or Traversable.
@@ -555,8 +560,6 @@ class PreOrderAnalysisVisitor extends ScopeVisitor
      */
     public function visitAssign(Node $node) : Context
     {
-        // In php 7.0, a **valid** parsed AST would be an ast\AST_LIST.
-        // However, --force-polyfill-parser will emit ast\AST_ARRAY.
         $var_node = $node->children['var'];
         if (Config::get_closest_target_php_version_id() < 70100 && $var_node instanceof Node && $var_node->kind === ast\AST_ARRAY) {
             $this->analyzeArrayAssignBackwardsCompatibility($var_node);
@@ -610,30 +613,23 @@ class PreOrderAnalysisVisitor extends ScopeVisitor
         // If there's a key, make a variable out of that too
         $key_node = $node->children['key'];
         if ($key_node instanceof Node) {
-            if ($key_node->kind == ast\AST_LIST) {
-                throw new NodeException(
-                    $node,
+            if ($key_node->kind === ast\AST_ARRAY) {
+                $this->emitIssue(
+                    Issue::InvalidNode,
+                    $key_node->lineno,
                     "Can't use list() as a key element - aborting"
                 );
-            }
-
-            $variable = Variable::fromNodeInContext(
-                $key_node,
-                $context,
-                $code_base,
-                false
-            );
-            if (!$expression_union_type->isEmpty()) {
+            } else {
                 // TODO: Support Traversable<Key, T> then return Key.
                 // If we see array<int,T> or array<string,T> and no other array types, we're reasonably sure the foreach key is an integer or a string, so set it.
                 // (Or if we see iterable<int,T>
-                $union_type_of_array_key = $expression_union_type->iterableKeyUnionType($code_base);
-                if ($union_type_of_array_key !== null) {
-                    $variable->setUnionType($union_type_of_array_key);
-                }
+                $context = (new AssignmentVisitor(
+                    $code_base,
+                    $context,
+                    $key_node,
+                    $expression_union_type->iterableKeyUnionType($code_base)
+                ))->__invoke($key_node);
             }
-
-            $context->addScopeVariable($variable);
         }
 
         // Note that we're not creating a new scope, just
@@ -655,8 +651,11 @@ class PreOrderAnalysisVisitor extends ScopeVisitor
             );
         }
         foreach ($union_type->getTypeSet() as $type) {
-            if ($type->asExpandedTypes($this->code_base)->hasTraversable()) {
-                continue;
+            try {
+                if ($type->asExpandedTypes($this->code_base)->hasTraversable()) {
+                    continue;
+                }
+            } catch (RecursionDepthException $_) {
             }
             if (!$type->isObjectWithKnownFQSEN()) {
                 continue;